refactor(@xen-orchestra/backups): _deltaVm => _incrementalVm
