+++ @xen-orchestra/backups/_incrementalVm.js
